# ARROW-9966: [Rust] Speedup kernels for sum,min,max by 10%-60%