--- a/rust/arrow/benches/aggregate_kernels.rs
+++ b/rust/arrow/benches/aggregate_kernels.rs
@@ -19,6 +19,7 @@
 extern crate criterion;
 use criterion::Criterion;
 
+use rand::Rng;
 use std::sync::Arc;
 
 extern crate arrow;
@@ -27,13 +28,15 @@ use arrow::array::*;
 use arrow::compute::kernels::aggregate::*;
 
 fn create_array(size: usize, with_nulls: bool) -> ArrayRef {
+    // use random numbers to avoid spurious compiler optimizations wrt to branching
+    let mut rng = rand::thread_rng();
     let mut builder = Float32Builder::new(size);
 
-    for i in 0..size {
-        if with_nulls && i % 2 == 0 {
+    for _ in 0..size {
+        if with_nulls && rng.gen::<f32>() > 0.5 {
             builder.append_null().unwrap();
         } else {
-            builder.append_value(1.0 + 1.0 * i as f32).unwrap();
+            builder.append_value(rng.gen()).unwrap();
         }
     }
     Arc::new(builder.finish())
@@ -44,14 +47,21 @@ fn bench_sum(arr_a: &ArrayRef) {
     criterion::black_box(sum(&arr_a).unwrap());
 }
 
+fn bench_min(arr_a: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
+    criterion::black_box(min(&arr_a).unwrap());
+}
+
 fn add_benchmark(c: &mut Criterion) {
     let arr_a = create_array(512, false);
 
     c.bench_function("sum 512", |b| b.iter(|| bench_sum(&arr_a)));
+    c.bench_function("min 512", |b| b.iter(|| bench_min(&arr_a)));
 
     let arr_a = create_array(512, true);
 
     c.bench_function("sum nulls 512", |b| b.iter(|| bench_sum(&arr_a)));
+    c.bench_function("min nulls 512", |b| b.iter(|| bench_min(&arr_a)));
 }
 
 criterion_group!(benches, add_benchmark);