--- a/rust/arrow/benches/arithmetic_kernels.rs
+++ b/rust/arrow/benches/arithmetic_kernels.rs
@@ -28,55 +28,61 @@ use arrow::compute::kernels::aggregate::*;
 use arrow::compute::kernels::arithmetic::*;
 use arrow::compute::kernels::limit::*;
 
-fn create_array(size: usize) -> Float32Array {
+fn create_array(size: usize) -> ArrayRef {
     let mut builder = Float32Builder::new(size);
-    for _i in 0..size {
-        builder.append_value(1.0).unwrap();
+    for i in 0..size {
+        builder.append_value(1.0 + 1.0 * i as f32).unwrap();
     }
-    builder.finish()
+    Arc::new(builder.finish())
 }
 
-fn bench_add(size: usize) {
-    let arr_a = create_array(size);
-    let arr_b = create_array(size);
-    criterion::black_box(add(&arr_a, &arr_b).unwrap());
+fn bench_add(arr_a: &ArrayRef, arr_b: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
+    let arr_b = arr_b.as_any().downcast_ref::<Float32Array>().unwrap();
+    criterion::black_box(add(arr_a, arr_b).unwrap());
 }
 
-fn bench_subtract(size: usize) {
-    let arr_a = create_array(size);
-    let arr_b = create_array(size);
+fn bench_subtract(arr_a: &ArrayRef, arr_b: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
+    let arr_b = arr_b.as_any().downcast_ref::<Float32Array>().unwrap();
     criterion::black_box(subtract(&arr_a, &arr_b).unwrap());
 }
 
-fn bench_multiply(size: usize) {
-    let arr_a = create_array(size);
-    let arr_b = create_array(size);
+fn bench_multiply(arr_a: &ArrayRef, arr_b: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
+    let arr_b = arr_b.as_any().downcast_ref::<Float32Array>().unwrap();
     criterion::black_box(multiply(&arr_a, &arr_b).unwrap());
 }
 
-fn bench_divide(size: usize) {
-    let arr_a = create_array(size);
-    let arr_b = create_array(size);
+fn bench_divide(arr_a: &ArrayRef, arr_b: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
+    let arr_b = arr_b.as_any().downcast_ref::<Float32Array>().unwrap();
     criterion::black_box(divide(&arr_a, &arr_b).unwrap());
 }
 
-fn bench_sum(size: usize) {
-    let arr_a = create_array(size);
+fn bench_sum(arr_a: &ArrayRef) {
+    let arr_a = arr_a.as_any().downcast_ref::<Float32Array>().unwrap();
     criterion::black_box(sum(&arr_a).unwrap());
 }
 
-fn bench_limit(size: usize, max: usize) {
-    let arr_a: ArrayRef = Arc::new(create_array(size));
-    criterion::black_box(limit(&arr_a, max).unwrap());
+fn bench_limit(arr_a: &ArrayRef, max: usize) {
+    criterion::black_box(limit(arr_a, max).unwrap());
 }
 
 fn add_benchmark(c: &mut Criterion) {
-    c.bench_function("add 512", |b| b.iter(|| bench_add(512)));
-    c.bench_function("subtract 512", |b| b.iter(|| bench_subtract(512)));
-    c.bench_function("multiply 512", |b| b.iter(|| bench_multiply(512)));
-    c.bench_function("divide 512", |b| b.iter(|| bench_divide(512)));
-    c.bench_function("sum 512", |b| b.iter(|| bench_sum(512)));
-    c.bench_function("limit 512, 512", |b| b.iter(|| bench_limit(512, 512)));
+    let arr_a = create_array(512);
+    let arr_b = create_array(512);
+
+    c.bench_function("add 512", |b| b.iter(|| bench_add(&arr_a, &arr_b)));
+    c.bench_function("subtract 512", |b| {
+        b.iter(|| bench_subtract(&arr_a, &arr_b))
+    });
+    c.bench_function("multiply 512", |b| {
+        b.iter(|| bench_multiply(&arr_a, &arr_b))
+    });
+    c.bench_function("divide 512", |b| b.iter(|| bench_divide(&arr_a, &arr_b)));
+    c.bench_function("sum 512", |b| b.iter(|| bench_sum(&arr_a)));
+    c.bench_function("limit 512, 512", |b| b.iter(|| bench_limit(&arr_a, 512)));
 }
 
 criterion_group!(benches, add_benchmark);